factor out ClassView and SlotView into separate modules
